Migration: added Team Leader Role

+++ db/migrate/20100407173131_add_stats_permissions.rb
@@ -34,15 +34,18 @@ class AddStatsPermissions < ActiveRecord::Migration
 
   def self.down
     NEW_MINISTRY_ROLE_PERMISSIONS.each do |mrp|
-      permission_id = Permission.find(:first, :conditions => { :controller => mrp[:controller], :action => mrp[:action] } ).id
+      p = Permission.find(:first, :conditions => { :controller => mrp[:controller], :action => mrp[:action] } )
+      if p
+        permission_id = p.id
 
-      p = MinistryRolePermission.find(:first, :conditions => {:permission_id => permission_id, :ministry_role_id => mrp[:ministry_role_id] } )
-      p.destroy
+        mr = MinistryRolePermission.find(:first, :conditions => {:permission_id => permission_id, :ministry_role_id => mrp[:ministry_role_id] } )
+        mr.destroy if mr
+      end
     end
 
     NEW_PERMISSIONS.each do |permission|
       p = Permission.find(:first, :conditions => { :controller => permission[:controller], :action => permission[:action] })
-      p.destroy
+      p.destroy if p
     end
   end
 end
